# src/api/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ErrorResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IVMResponse": {
				"properties": {
					"cpuUsage": {
						"type": "string"
					},
					"ramUsagePercent": {
						"type": "string"
					},
					"ramUsage": {
						"type": "string"
					},
					"totalRam": {
						"type": "string"
					},
					"freeRam": {
						"type": "string"
					},
					"totalDisk": {
						"type": "string"
					},
					"freeDisk": {
						"type": "string"
					},
					"diskUsage": {
						"type": "string"
					},
					"diskUsagePercent": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"staticId": {
						"type": "string"
					},
					"dynamicId": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"macAddress": {
						"type": "string"
					},
					"ipAddress": {
						"type": "string"
					},
					"port": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ActionResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					},
					"key": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "spinal-organ-health",
		"version": "1.0.1",
		"description": "API for Spinal Organ Health Monitoring",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "SpinalCom",
			"email": "developers@spinalcom.com"
		}
	},
	"paths": {
...
		"/monitoring/{vmKey}": {
			"get": {
				"operationId": "GetVirtualMachine",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IVMResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/ram/value": {
			"get": {
				"operationId": "GetPm2RamHistoryValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointValueResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/ram/timeseries": {
			"get": {
				"operationId": "GetPm2RamHistoryTimeseries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointTimeSeriesResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "endTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/cpu/value": {
			"get": {
				"operationId": "GetPm2CpuHistoryValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointValueResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/cpu/timeseries": {
			"get": {
				"operationId": "GetPm2CpuHistoryTimeseries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointTimeSeriesResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "endTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/heap_size/value": {
			"get": {
				"operationId": "GetPm2HeapSizeHistoryValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointValueResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/heap_size/timeseries": {
			"get": {
				"operationId": "GetPm2HeapSizeHistoryTimeseries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointTimeSeriesResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "endTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/heap_usage/value": {
			"get": {
				"operationId": "GetPm2HeapUsageHistoryValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointValueResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/heap_usage/timeseries": {
			"get": {
				"operationId": "GetPm2HeapUsageHistoryTimeseries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointTimeSeriesResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "endTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/heap_used_size/value": {
			"get": {
				"operationId": "GetPm2HeapUsedSizeHistoryValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointValueResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/heap_used_size/timeseries": {
			"get": {
				"operationId": "GetPm2HeapUsedSizeHistoryTimeseries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointTimeSeriesResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "endTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/reboot/value": {
			"get": {
				"operationId": "GetPm2RebootHistoryValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointValueResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/reboot/timeseries": {
			"get": {
				"operationId": "GetPm2RebootHistoryTimeseries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointTimeSeriesResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "endTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/errored/value": {
			"get": {
				"operationId": "GetPm2ErroredHistoryValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointValueResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/errored/timeseries": {
			"get": {
				"operationId": "GetPm2ErroredHistoryTimeseries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/EndpointTimeSeriesResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "endTime",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/monitoring/endpoints/{vmKey}/pm2/{pm2Key}/{endpoint}/timeseries/maxDay": {
			"post": {
				"operationId": "UpdatePm2EndpointMaxDay",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ActionResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "pm2Key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "endpoint",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMaxDayBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}